+++ src/main/java/fll/web/WebUtils.java
@@ -390,6 +390,12 @@ public final class WebUtils {
   public static final String SLIDESHOW_PATH = "slideshow";
 
   /**
+   * The path to the custom images relattive to the root of the web application.
+   * Does not have a leading or trailing slash.
+   */
+  public static final String CUSTOM_PATH = "custom";
+
+  /**
    * Get the sponsor logo filenames.
    *
    * @param application used to get the absolute path of the sponsor logos